+++ src/main/java/com/ss/ita/google/ui/runner/TestRunnerForGoogle.java
@@ -1,4 +1,4 @@
-package com.ss.ita.rozetka.ui.pages.runner;
+package com.ss.ita.google.ui.runner;
 
 import com.codeborne.selenide.Configuration;
 import org.testng.annotations.BeforeClass;
@@ -6,12 +6,12 @@ import org.testng.annotations.BeforeClass;
 import static com.codeborne.selenide.Configuration.*;
 import static com.codeborne.selenide.Selenide.*;
 
-public class TestRunnerForRozetka {
+public class TestRunnerForGoogle {
 
     @BeforeClass
     protected void setBrowser() {
         browser = "chrome";
         Configuration.browserSize = "1920x1080";
-        open("https://rozetka.com.ua");
+        open("https://google.com");
     }
 }
