+++ src/main/java/com/ss/ita/rozetka/ui/runner/TestRunnerForRozetka.java
@@ -1,4 +1,4 @@
-package com.ss.ita.google.ui.pages.runner;
+package com.ss.ita.rozetka.ui.runner;
 
 import com.codeborne.selenide.Configuration;
 import org.testng.annotations.BeforeClass;
@@ -6,12 +6,12 @@ import org.testng.annotations.BeforeClass;
 import static com.codeborne.selenide.Configuration.*;
 import static com.codeborne.selenide.Selenide.*;
 
-public class TestRunnerForGoogle {
+public class TestRunnerForRozetka {
 
     @BeforeClass
     protected void setBrowser() {
         browser = "chrome";
         Configuration.browserSize = "1920x1080";
-        open("https://google.com");
+        open("https://rozetka.com.ua");
     }
 }
