+++ src/test/java/com/ss/ita/google/HideHomePageLogoTest.java
@@ -1,8 +1,8 @@
 package com.ss.ita.google;
 
 import com.ss.ita.google.ui.pages.HomePage;
-import com.ss.ita.google.ui.pages.runner.TestRunnerForGoogle;
-import static com.ss.ita.google.ui.pages.util.WebElementUtil.*;
+import com.ss.ita.google.ui.runner.TestRunnerForGoogle;
+import static com.ss.ita.google.ui.util.WebElementUtil.*;
 import org.testng.Assert;
 import org.testng.annotations.Test;
