+++ src/test/java/com/ss/ita/google/ResultLinkTest.java
@@ -2,12 +2,11 @@ package com.ss.ita.google;
 
 import com.ss.ita.google.ui.pages.HomePage;
 import com.ss.ita.google.ui.pages.SearchResultPage;
-import com.ss.ita.google.ui.pages.runner.TestRunnerForGoogle;
 import org.testng.annotations.Test;
 
 import static org.testng.Assert.assertEquals;
 
-public class ResultLinkTest extends TestRunnerForGoogle {
+public class ResultLinkTest {
 
     @Test
     public void verifyOpenResultLink() {
