+++ src/test/java/com/ss/ita/google/SearchAndGoBackHomePageTest.java
@@ -2,7 +2,7 @@ package com.ss.ita.google;
 
 import com.ss.ita.google.ui.pages.HomePage;
 
-import com.ss.ita.google.ui.pages.runner.TestRunnerForGoogle;
+import com.ss.ita.google.ui.runner.TestRunnerForGoogle;
 import org.testng.annotations.Test;
 import org.testng.asserts.SoftAssert;
