+++ src/test/java/com/ss/ita/google/SearchLinksContainsWordTest.java
@@ -2,7 +2,7 @@ package com.ss.ita.google;
 
 import com.ss.ita.google.ui.pages.HomePage;
 import com.ss.ita.google.ui.pages.SearchResultPage;
-import com.ss.ita.google.ui.pages.runner.TestRunnerForGoogle;
+import com.ss.ita.google.ui.runner.TestRunnerForGoogle;
 import org.testng.annotations.Test;
 
 import java.util.List;
